Move cypress & netscript tests to ./test subfolder
