+++ cypress.json
@@ -1,10 +1,14 @@
 {
   "baseUrl": "http://localhost:8000",
-  "fixturesFolder": false,
   "trashAssetsBeforeRuns": true,
-  "screenshotsFolder": ".cypress/screenshots",
-  "videosFolder": ".cypress/videos",
   "videoUploadOnPasses": false,
   "viewportWidth": 1980,
-  "viewportHeight": 1080
+  "viewportHeight": 1080,
+  "fixturesFolder": "test/cypress/fixtures",
+  "integrationFolder": "test/cypress/integration",
+  "pluginsFile": "test/cypress/plugins/index.js",
+  "supportFile": "test/cypress/support/index.js",
+  "screenshotsFolder": ".cypress/screenshots",
+  "videosFolder": ".cypress/videos",
+  "downloadsFolder": ".cypress/downloads"
 }
