+++ test/cypress/integration/netscript.spec.ts
