+++ test/jest/Script/Script.test.ts
@@ -1,8 +1,8 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { jest, describe, expect, test } from "@jest/globals";
 
-import { Script } from "../../src/Script/Script";
-import { Player } from "../../src/Player";
+import { Script } from "../../../src/Script/Script";
+import { Player } from "../../../src/Player";
 
 jest.mock(`!!raw-loader!../NetscriptDefinitions.d.ts`, () => "", {
     virtual: true,
