+++ test/jest/Terminal/Directory.test.js
@@ -1,6 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { jest, describe, expect, test } from '@jest/globals'
-import * as dirHelpers from "../../src/Terminal/DirectoryHelpers";
+import * as dirHelpers from "../../../src/Terminal/DirectoryHelpers";
 
 describe("Terminal Directory Tests", function () {
   describe("removeLeadingSlash()", function () {
