+++ test/jest/Terminal/determineAllPossibilitiesForTabCompletion.test.ts
@@ -2,12 +2,12 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { jest, describe, expect, test } from "@jest/globals";
 
-import { Player } from "../../src/Player";
-import { determineAllPossibilitiesForTabCompletion } from "../../src/Terminal/determineAllPossibilitiesForTabCompletion";
-import { Server } from "../../src/Server/Server";
-import { AddToAllServers, prestigeAllServers } from "../../src/Server/AllServers";
-import { LocationName } from "../../src/Locations/data/LocationNames";
-import { CodingContract } from "../../src/CodingContracts";
+import { Player } from "../../../src/Player";
+import { determineAllPossibilitiesForTabCompletion } from "../../../src/Terminal/determineAllPossibilitiesForTabCompletion";
+import { Server } from "../../../src/Server/Server";
+import { AddToAllServers, prestigeAllServers } from "../../../src/Server/AllServers";
+import { LocationName } from "../../../src/Locations/data/LocationNames";
+import { CodingContract } from "../../../src/CodingContracts";
 
 jest.mock(`!!raw-loader!../NetscriptDefinitions.d.ts`, () => "", {
   virtual: true,
