+++ test/netscript/tb_basic.script
