+++ test/netscript/tb_foo.script
