+++ test/netscript/tb_functions.script
