+++ test/netscript/tb_multiarray.script
