+++ test/netscript/tb_start.script
